+++ e2e-tests/insurance/cypress/e2e/journeys/export-contract/agent-service/save-and-back.spec.js
@@ -94,11 +94,8 @@ context('Insurance - Export contract - Agent service - Save and go back', () => 
 
         cy.startInsuranceExportContractSection({});
 
-        // TODO: EMS-3665 - increment the value in command below instead.
-        cy.completeAndSubmitHowWasTheContractAwardedForm({});
-
-        // go through 4 export contract forms.
-        cy.clickSubmitButtonMultipleTimes({ count: 4 });
+        // go through 5 export contract forms.
+        cy.clickSubmitButtonMultipleTimes({ count: 5 });
 
         cy.assertAgentServiceFieldValues({});
       });
@@ -124,11 +121,8 @@ context('Insurance - Export contract - Agent service - Save and go back', () => 
 
         cy.startInsuranceExportContractSection({});
 
-        // TODO: EMS-3665 - increment the value in command below instead.
-        cy.completeAndSubmitHowWasTheContractAwardedForm({});
-
-        // go through 4 export contract forms.
-        cy.clickSubmitButtonMultipleTimes({ count: 4 });
+        // go through 5 export contract forms.
+        cy.clickSubmitButtonMultipleTimes({ count: 5 });
 
         cy.assertAgentServiceFieldValues({ agentIsCharging: true });
       });