+++ src/ui/server/helpers/task-list/generate-groups-and-tasks/index.test.ts
@@ -27,6 +27,7 @@ describe('server/helpers/task-list/generate-groups-and-tasks', () => {
           charge: { method: agentChargeMethod },
         },
       },
+      awardMethod: { id: awardMethodId },
     },
     nominatedLossPayee: {
       isAppointed: isAppointingLossPayee,
@@ -56,6 +57,10 @@ describe('server/helpers/task-list/generate-groups-and-tasks', () => {
       exporterHasPreviousCreditInsuranceWithBuyer,
       totalContractValueOverThreshold,
       attemptedPrivateMarketCover,
+      isUsingAgent,
+      agentIsCharging,
+      agentChargeMethod,
+      awardMethodId,
     );
 
     const initialChecks = {
@@ -87,6 +92,7 @@ describe('server/helpers/task-list/generate-groups-and-tasks', () => {
         isUsingAgent,
         agentIsCharging,
         agentChargeMethod,
+        awardMethodId,
       }),
     };